+++ index.html
@@ -226,6 +226,6 @@
       </div>
     </section>
   </template>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
